Crate featomic

Re-exports

pub use self::systems::System;

pub use self::systems::SimpleSystem;

pub use types::*;

Modules

calculators

labels

math

cbindgen:ignore

systems

types

This module provides 3D vectors and matrix to be used in all other modules.

Structs

CalculationOptions

Parameters specific to a single call to compute

Calculator

Enums

Error

LabelsSelection

Rules to select labels (either samples or properties) on which the user wants to run a calculation