featomic::calculators
Module lode

Structs

LodeRadialIntegralGto

Implementation of the LODE radial integral for GTO radial basis and Gaussian atomic density.

LodeRadialIntegralSpline

LodeRadialIntegralSpline allows to evaluate another radial integral implementation using cubic Hermit spline.

LodeSphericalExpansion

The actual calculator used to compute LODE spherical expansion coefficients

LodeSphericalExpansionParameters

Parameters for LODE spherical expansion calculator.

Traits

LodeRadialIntegral

A LodeRadialIntegral computes the LODE radial integral for all radial basis functions and a single spherical harmonic l channel