featomic::calculators::lode
Struct LodeRadialIntegralSpline

pub struct LodeRadialIntegralSpline { /* private fields */ }

LodeRadialIntegralSpline allows to evaluate another radial integral implementation using cubic Hermit spline.

This can be much faster than using the actual radial integral implementation.

Implementations

impl LodeRadialIntegralSpline

pub fn with_accuracy( radial_integral: impl LodeRadialIntegral, density: DensityKind, k_cutoff: f64, accuracy: f64, with_center_contribution: bool, ) -> Result<LodeRadialIntegralSpline, Error>

Create a new LodeRadialIntegralSpline taking values from the given radial_integral. Points are added to the spline until the requested accuracy is reached. We consider that the accuracy is reached when either the mean absolute error or the mean relative error gets below the accuracy threshold.

pub fn from_tabulated( tabulated: LodeTabulated, density: DensityKind, ) -> LodeRadialIntegralSpline

Create a new LodeRadialIntegralSpline with user-defined spline points.

The density/tabulated.center_contribution are assumed to match each other