featomic::calculators::lode
Trait LodeRadialIntegral

pub trait LodeRadialIntegral: RefUnwindSafe + Send {
    // Required methods
    fn compute(
        &self,
        k_norm: f64,
        values: ArrayViewMut1<'_, f64>,
        gradients: Option<ArrayViewMut1<'_, f64>>,
    );
    fn size(&self) -> usize;
    fn get_center_contribution(
        &self,
        density: DensityKind,
    ) -> Result<Array1<f64>, Error>;
}

A LodeRadialIntegral computes the LODE radial integral for all radial basis functions and a single spherical harmonic l channel

See equations 5 to 8 of this paper for mor information on the radial integral.

std::panic::RefUnwindSafe is a required super-trait to enable passing radial integrals across the C API. Send is a required super-trait to enable passing radial integrals between threads.

Required Methods

fn compute( &self, k_norm: f64, values: ArrayViewMut1<'_, f64>, gradients: Option<ArrayViewMut1<'_, f64>>, )

Compute the LODE radial integral for a single k-vector norm and store the resulting data in the values array. If gradients is Some, also compute and store gradients there.

fn size(&self) -> usize

Get how many basis functions are part of this integral. This is the shape to use for the values and gradients parameters to compute.

fn get_center_contribution( &self, density: DensityKind, ) -> Result<Array1<f64>, Error>

Compute the contribution of the central atom to the final <n l m> coefficients. By symmetry, only l=0 is non-zero, so this function returns a 1-D array containing the different <n 0 0> coefficients.

This function differs from the rest of LODE calculation because it goes straight from atom => n l m, without using k-space projection in the middle.