featomic::calculators

Module soap

Structs

Cutoff

Definition of a local environment for SOAP calculations

PowerSpectrumParameters

Parameters for SOAP power spectrum calculator.

RadialSpectrumParameters

Parameters for the SOAP radial spectrum calculator.

SoapPowerSpectrum

Calculator implementing the Smooth Overlap of Atomic Position (SOAP) power spectrum representation of atomistic systems.

SoapRadialIntegralCacheByAngular

Store all SoapRadialIntegralCache for different angular channels

SoapRadialSpectrum

Calculator implementing the Radial spectrum representation of atomistic systems.

SphericalExpansion

The actual calculator used to compute SOAP spherical expansion coefficients

SphericalExpansionByPair

The actual calculator used to compute spherical expansion pair-by-pair

SphericalExpansionParameters

Parameters for spherical expansion calculator.

Enums

Smoothing

Possible values for the smoothing cutoff function

Traits

SoapRadialIntegral

A SoapRadialIntegral computes the SOAP radial integral for all radial basis functions and a single spherical harmonic l channel