featomic::calculators::soap
Trait SoapRadialIntegral

pub trait SoapRadialIntegral: RefUnwindSafe + Send {
    // Required methods
    fn compute(
        &self,
        distance: f64,
        values: ArrayViewMut1<'_, f64>,
        gradients: Option<ArrayViewMut1<'_, f64>>,
    );
    fn size(&self) -> usize;
}

A SoapRadialIntegral computes the SOAP radial integral for all radial basis functions and a single spherical harmonic l channel

See equations 5 to 8 of this paper for mor information on the radial integral.

std::panic::RefUnwindSafe is a required super-trait to enable passing radial integrals across the C API. Send is a required super-trait to enable passing radial integrals between threads.

Required Methods

fn compute( &self, distance: f64, values: ArrayViewMut1<'_, f64>, gradients: Option<ArrayViewMut1<'_, f64>>, )

Compute the radial integral for a single distance between two atoms and store the resulting data in the values array. If gradients is Some, also compute and store gradients there.

The radial integral $I_{nl}$ is defined as “the non-spherical harmonics part of the spherical expansion”. Depending on the atomic density, different expressions can be used.

For a delta density, the radial integral is simply the radial basis function $R_{nl}$ evaluated at the pair distance:

$$ I_{nl}(r_{ij}) = R_{nl}(r_{ij}) $$

For a Gaussian atomic density with a width of $\sigma$, the radial integral reduces to:

$$ I_{nl}(r_{ij}) = \frac{4\pi}{(\pi \sigma^2)^{3/4}} e^{-\frac{r_{ij}^2}{2\sigma^2}} \int_0^\infty \mathrm{d}r r^2 R_{nl}(r) e^{-\frac{r^2}{2\sigma^2}} i_l\left(\frac{rr_{ij}}{\sigma^2}\right) $$

where $i_l$ is the modified spherical Bessel function of the first kind of order $l$.

Finally, for an arbitrary spherically symmetric atomic density g(r), the radial integral is

$$ I_{nl}(r_{ij}) = 2\pi \int_0^\infty \mathrm{d}r r^2 R_{nl}(r) \int_{-1}^1 \mathrm{d}u P_l(u) g(\sqrt{r^2+r_{ij}^2-2rr_{ij}u}) $$

where $P_l$ is the l-th Legendre polynomial.

fn size(&self) -> usize

Get how many basis functions are part of this integral. This is the shape to use for the values and gradients parameters to compute.