featomic::calculators
Trait CalculatorBase

pub trait CalculatorBase: RefUnwindSafe {
    // Required methods
    fn name(&self) -> String;
    fn parameters(&self) -> String;
    fn cutoffs(&self) -> &[f64];
    fn keys(&self, systems: &mut [Box<dyn System>]) -> Result<Labels, Error>;
    fn sample_names(&self) -> Vec<&str>;
    fn samples(
        &self,
        keys: &Labels,
        systems: &mut [Box<dyn System>],
    ) -> Result<Vec<Labels>, Error>;
    fn supports_gradient(&self, parameter: &str) -> bool;
    fn positions_gradient_samples(
        &self,
        keys: &Labels,
        samples: &[Labels],
        systems: &mut [Box<dyn System>],
    ) -> Result<Vec<Labels>, Error>;
    fn components(&self, keys: &Labels) -> Vec<Vec<Labels>>;
    fn property_names(&self) -> Vec<&str>;
    fn properties(&self, keys: &Labels) -> Vec<Labels>;
    fn compute(
        &mut self,
        systems: &mut [Box<dyn System>],
        descriptor: &mut TensorMap,
    ) -> Result<(), Error>;
}

The CalculatorBase trait is the interface shared by all calculator implementations; and used by crate::Calculator to run the calculation.

This should not be used directly by end users, who should use the facilities in crate::Calculator instead.

std::panic::RefUnwindSafe is a required super-trait to enable passing calculators across the C API.

Required Methods

fn name(&self) -> String

Get the name of this Calculator

fn parameters(&self) -> String

Get the parameters used to create this Calculator as a JSON string

fn cutoffs(&self) -> &[f64]

Get the all radial cutoffs used by this Calculator’s neighbors lists (which can be an empty list)

fn keys(&self, systems: &mut [Box<dyn System>]) -> Result<Labels, Error>

Get the set of keys for this calculator and the given systems

fn sample_names(&self) -> Vec<&str>

Get the names used for sample labels by this calculator

fn samples( &self, keys: &Labels, systems: &mut [Box<dyn System>], ) -> Result<Vec<Labels>, Error>

Get the full list of samples this calculator would create for the given systems. This function should return one set of samples for each key.

fn supports_gradient(&self, parameter: &str) -> bool

Can this calculator compute gradients with respect to the parameter? Right now, parameter can be either "positions", "strain" or "cell".

fn positions_gradient_samples( &self, keys: &Labels, samples: &[Labels], systems: &mut [Box<dyn System>], ) -> Result<Vec<Labels>, Error>

Get the samples for gradients with respect to positions, corresponding the given values samples.

The samples slice contains one set of samples for each key.

If the gradients with respect to positions are not available, this function should return an error.

fn components(&self, keys: &Labels) -> Vec<Vec<Labels>>

Get the components this calculator computes for each key.

fn property_names(&self) -> Vec<&str>

Get the names used for property labels by this calculator

fn properties(&self, keys: &Labels) -> Vec<Labels>

Get the properties this calculator computes for each key.

fn compute( &mut self, systems: &mut [Box<dyn System>], descriptor: &mut TensorMap, ) -> Result<(), Error>

Actually run the calculation.

This function is given a pre-allocated descriptor, filled with zeros. The samples/properties in each blocks might not match the values returned by CalculatorBase::samples and CalculatorBase::properties: instead they will only contain the values that where requested by the end user.

Gradients (with respect to positions or cell) are allocated in each block if they are supported according to CalculatorBase::supports_gradient, and the users requested them as part of the calculation options.