featomic
Module labels

Structs

AllTypesPairsKeys

Compute a set of keys with two variables: the central atom type and a all neighbor atom types within the whole system.

AtomCenteredSamples

SampleBuilder for atom-centered representation. This will create one sample for each atom, optionally filtering on the central atom type. The sample names are (“system”, “atom”).

CenterSingleNeighborsTypesKeys

Compute a set of keys with two variables: the central atom type and a single neighbor atom type within a cutoff around the central atom.

CenterTwoNeighborsTypesKeys

Compute a set of keys with three variables: the central atom type and two neighbor atom types.

CenterTypesKeys

Compute a set of keys with a single variable, the central atom type.

LongRangeSamplesPerAtom

Samples builder for long-range (i.e. infinite cutoff), per atom, two bodies representation.

Enums

AtomicTypeFilter

Atomic type filters to be used when building samples and gradient sample

Traits

KeysBuilder

Common interface to create a set of metatensor’s TensorMap keys from systems

SamplesBuilder

Abstraction over the different kinds of samples used in featomic.