featomic::labels
Trait SamplesBuilder

pub trait SamplesBuilder {
    // Required methods
    fn sample_names() -> Vec<&'static str>;
    fn samples(&self, systems: &mut [Box<dyn System>]) -> Result<Labels, Error>;
    fn gradients_for(
        &self,
        systems: &mut [Box<dyn System>],
        samples: &Labels,
    ) -> Result<Labels, Error>;
}

Abstraction over the different kinds of samples used in featomic.

Different implementations of this trait correspond to different types of samples (for example one sample for each system; or one sample for each pair, etc.)

Each implementation must be able to generate samples from a list of systems, and can optionally implement samples for gradients calculation.

Required Methods

fn sample_names() -> Vec<&'static str>

Get the names used by the samples

fn samples(&self, systems: &mut [Box<dyn System>]) -> Result<Labels, Error>

Create Labels containing all the samples corresponding to the given list of systems.

fn gradients_for( &self, systems: &mut [Box<dyn System>], samples: &Labels, ) -> Result<Labels, Error>

Create a set of Labels containing the gradient samples corresponding to the given samples in the given systems; and only these.

Object Safety

This trait is not object safe.