featomic

Module calculators

Re-exports

pub use self::soap::SphericalExpansionByPair;

pub use self::soap::SphericalExpansionParameters;

pub use self::soap::SphericalExpansion;

pub use self::soap::SoapRadialSpectrum;

pub use self::soap::RadialSpectrumParameters;

pub use self::soap::SoapPowerSpectrum;

pub use self::soap::PowerSpectrumParameters;

pub use self::lode::LodeSphericalExpansion;

pub use self::lode::LodeSphericalExpansionParameters;

Modules

lode

soap

Structs

AtomicComposition

An atomic composition calculator for obtaining the stoichiometric information.

Density

Definition of the (atomic) density to expand on a basis

NeighborList

This calculator computes the neighbor list for a given spherical cutoff, and returns the list of distance vectors between all pairs of atoms strictly inside the cutoff.

SortedDistances

Sorted distances vector representation of an atomic environment.

TensorProductBasis

Information about “tensor product” spherical expansion basis functions

Enums

DensityKind

Different available kinds of atomic density to use in featomic

DensityScaling

Implemented options for radial scaling of the atomic density around an atom

LodeRadialBasis

The different kinds of radial basis supported LODE calculators

SoapRadialBasis

The different kinds of radial basis supported by SOAP calculators

SphericalExpansionBasis

Possible Basis functions to use for the SOAP or LODE spherical expansion.

Traits

CalculatorBase

The CalculatorBase trait is the interface shared by all calculator implementations; and used by crate::Calculator to run the calculation.